pyproject.toml
[project]
name = "job-cost-tool"
version = "0.1.0"
requires-python = "~=3.12.0"
dependencies = [
  "pdfplumber",
  "openpyxl",
  "fastapi",
  "httpx",
  "python-multipart",
  "uvicorn",
  "psycopg[binary]",
  "vercel",
  "python-dotenv",
]
